requirements.txt
amqp==1.4.6
anyjson==0.3.3
arrow==0.5.4
billiard==3.3.0.20
celery==3.1.18
click==5.1
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-bootstrap3==6.2.2
django-crispy-forms==1.5.2
django-formtools==1.0
django-gravatar2==1.3.0
django-haystack==2.4.0
django-js-reverse==0.5.1
django-redis==4.1.0
django-redisboard==1.2.0
django-registration-bootstrap==0.2.0
django-registration-redux==1.2
django-th==0.11.0
django-toolbelt==0.0.1
djangorestframework==3.2.4
dropbox==3.34
feedparser==5.1.3
github3.py==1.0.0a2
gunicorn==19.3.0
httplib2==0.9.1
kombu==3.0.26
msgpack-python==0.4.6
oauthlib==1.0.3
ply==3.4
pocket==0.3.5
psycopg2==2.6.1
py==1.4.30
py-trello==0.4.3
PyJWT==1.4.0
pytest==2.8.0
python-dateutil==2.4.2
python-social-auth==0.2.13
python3-openid==3.0.7
pytidylib6==0.2.2
pytz==2015.6
readability-api==1.0.0
redis==2.10.3
requests==2.7.0
requests-oauthlib==0.5.0
rq==0.5.5
six==1.9.0
slimit==0.8.1
static3==0.6.1
tweepy==3.4.0
twython==3.2.0
uritemplate.py==0.3.0
urllib3==1.12
whitenoise==2.0.4
